requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
greenlet==3.4.0
idna==3.11
playwright==1.58.0
pyee==13.0.1
pyodbc==5.3.0
pyotp==2.9.0
python-dotenv==1.2.2
RapidFuzz==3.14.5
rarfile==4.2
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
